moving render for svg to another file
and requiring it in renderer/svg
