+++ lib/browser.js
@@ -1,6 +1,6 @@
 var QRCode = require('./core/qrcode')
 var CanvasRenderer = require('./renderer/canvas')
-var SvgRenderer = require('./renderer/svg')
+var SvgRenderer = require('./renderer/svg-render.js')
 
 function renderCanvas (renderFunc, canvas, text, opts, cb) {
   var argsNum = arguments.length - 1
